requirements.txt
fastapi
uvicorn
httpx
openai
python-dotenv
azure-cosmos
azure.identity